+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/ImportOrderingRule.kt
@@ -169,13 +169,28 @@ public class ImportOrderingRule :
                     it.elementType == ElementType.IMPORT_DIRECTIVE ||
                         it.psi is PsiWhiteSpace && it.textLength > 1 // also collect empty lines, that are represented as "\n\n"
                 }
+
+                val autoCorrectDuplicateImports = imports
+                    .filter { it.psi !is PsiWhiteSpace }
+                    .groupingBy { it.text }
+                    .eachCount()
+                    .filterValues { it > 1 }
+                    .keys
+                    .map {
+                        emit(
+                            node.startOffset,
+                            "Duplicate '$it' found",
+                            true
+                        )
+                    }.any()
+                val deduplicatedImports = imports.distinctBy { it.text } // distinguish by import path including alias
+
                 val hasComments = children.find { it.elementType == ElementType.BLOCK_COMMENT || it.elementType == ElementType.EOL_COMMENT } != null
-                val sortedImports = imports
+                val sortedImports = deduplicatedImports
                     .asSequence()
                     .filter { it.psi !is PsiWhiteSpace } // sorter expects KtImportDirective, whitespaces are inserted afterwards
                     .map { it.psi as KtImportDirective }
                     .sortedWith(importSorter)
-                    .distinctBy { if (it.aliasName != null) it.text.substringBeforeLast(it.aliasName!!) else it.text } // distinguish by import path w/o aliases
                     .map { it.node } // transform back to ASTNode in order to operate over its method (addChild)
 
                 // insert blank lines wherever needed
@@ -201,19 +216,24 @@ public class ImportOrderingRule :
                     return@fold current
                 }
 
-                val canAutoCorrect = !hasComments
-                if (!importsAreEqual(imports, sortedImportsWithSpaces) || (hasTooMuchWhitespace(children) && !isCustomLayout())) {
-                    val additionalMessage = if (!canAutoCorrect) {
-                        " -- no autocorrection due to comments in the import list"
-                    } else {
-                        ""
-                    }
+                if (hasComments) {
                     emit(
                         node.startOffset,
-                        "${errorMessages.getOrDefault(importsLayout, CUSTOM_ERROR_MESSAGE)}$additionalMessage",
-                        canAutoCorrect
+                        errorMessages.getOrDefault(importsLayout, CUSTOM_ERROR_MESSAGE) +
+                            " -- no autocorrection due to comments in the import list",
+                        false
                     )
-                    if (autoCorrect && canAutoCorrect) {
+                } else {
+                    val autoCorrectWhitespace = hasTooMuchWhitespace(children) && !isCustomLayout()
+                    val autoCorrectSortOrder = !importsAreEqual(deduplicatedImports, sortedImportsWithSpaces)
+                    if (autoCorrectSortOrder || autoCorrectWhitespace) {
+                        emit(
+                            node.startOffset,
+                            errorMessages.getOrDefault(importsLayout, CUSTOM_ERROR_MESSAGE),
+                            true
+                        )
+                    }
+                    if (autoCorrect && (autoCorrectDuplicateImports || autoCorrectSortOrder || autoCorrectWhitespace)) {
                         node.removeRange(node.firstChildNode, node.lastChildNode.treeNext)
                         sortedImportsWithSpaces.reduce { current, next ->
                             node.addChild(current, null)