+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/NoUnusedImportsRuleTest.kt
@@ -860,4 +860,18 @@ class NoUnusedImportsRuleTest {
             )
         )
     }
+
+    @Test
+    fun `Issue 1243 - Format should not remove import`() {
+        val code =
+            """
+            import foo.Bar as Bar1
+            import foo.Bar as Bar2
+
+            val bar1 = Bar1()
+            val bar2 = Bar2()
+            """.trimIndent()
+        assertThat(NoUnusedImportsRule().lint(code)).isEmpty()
+        assertThat(NoUnusedImportsRule().format(code)).isEqualTo(code)
+    }
 }